blockNumber

+++ ios/Classes/SwiftNfcManagerPlugin.swift
@@ -292,7 +292,7 @@ public class SwiftNfcManagerPlugin: NSObject, FlutterPlugin {
             result(FlutterError(code: "not_found", message: "Tag is not found.", details: nil))
             return
         }
-        connectedTech.readSingleBlock(requestFlags:requestFlags,blockNumber:UInt8(blockNumber)) { data, error in
+        connectedTech.readSingleBlock(requestFlags:requestFlags,blockNumber:blockNumber) { data, error in
             if let error = error {
                 result(error.toFlutterError())
                 return
